Rewrite entire site generation to work with the Jekyll framework.
This large commit switches from a everthing-is-hardcoded approach to
using the Jekyll framework to generate the site structure from
data and the layouts. This has the advantage that layouts and partials
only have to be written once instead of being copied into every subpage.

It also allows for faster data entry since the content can be defined
in a more descriptive fashion.